Update README.md
In today’s digital world, where patience is limited and users expect instant results, the performance of a website is no longer just an IT concern. It’s a customer service necessity. A slow-loading website isn’t just frustrating for users; it’s a potential deal breaker. It could mean lost sales, reduced trust, and a tarnished brand reputation.